Chore: move Filters types to types dir
