+++ src/filters/stores/useFiltersStore.ts
@@ -1,29 +1,7 @@
 import { toast } from "@/ui/components/use-toast";
 import { httpClient } from "@/lib/httpClient";
 import { create } from "zustand";
-
-
-// TODO: move to ../types
-export interface Filter {
-    id: string;
-    name: string;
-    criteria: string;
-    threshold?: number;
-    strictness?: number;
-    category: string;
-    created_at: string;
-    updated_at?: string;
-    is_custom?: boolean;
-}
-
-export interface FilterCreateRequest {
-    name: string;
-    criteria: string;
-    threshold: number;
-    strictness: number;
-    category?: string;
-    template?: string;
-}
+import { Filter, FilterCreateRequest } from "../types";
 
 export interface FiltersStore {
     systemFilters: Filter[];
