+++ src/filters/types/index.ts
@@ -1,0 +1,20 @@
+export interface Filter {
+    id: string;
+    name: string;
+    criteria: string;
+    threshold?: number;
+    strictness?: number;
+    category: string;
+    created_at: string;
+    updated_at?: string;
+    is_custom?: boolean;
+}
+
+export interface FilterCreateRequest {
+    name: string;
+    criteria: string;
+    threshold: number;
+    strictness: number;
+    category?: string;
+    template?: string;
+}
